tools/monarchmoney/requirements.txt
aiohttp>=3.8.4
gql>=4.0
oathtool>=2.3.1
pyyaml>=6.0
cryptography>=41.0.0
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0
pydantic>=2.0.0
mcp>=0.9.0
